testsuite: Consolidate afp_ls as a command in afparg
- Adds FPEnumerate <dir> as an afparg command
- Moves test300 from afp_ls to spectest
- Completely remove afp_ls as a separate executable
